feat(architecture): introduce minimal protocol package for client-server communication
- Added @civjs/protocol package with minimal shared network types
- Defined action types and diplomatic state types for client-server protocol
- Created lightweight client-only nation types for UI display
- Updated client and server to depend on protocol package instead of shared
- Added architecture recommendation document outlining new minimal shared dependencies approach

This change reduces client bundle size significantly by removing heavy nation data from client and separating concerns between client and server. It establishes a clear contract for network communication and enables independent scaling and deployment of client and server.
