+++ packages/protocol/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@civjs/protocol",
+  "version": "1.0.0",
+  "description": "Minimal network protocol types for CivJS client-server communication",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "tsc",
+    "clean": "rm -rf dist"
+  },
+  "devDependencies": {
+    "typescript": "^5.0.0"
+  }
+}
