--- packages/shared/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@civjs/shared",
-  "version": "1.0.0",
-  "description": "Shared types and data for CivJS client and server",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist/**/*",
-    "src/**/*"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "clean": "rm -rf dist"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "typescript": "^5.0.0"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.js"
-    },
-    "./types/nations": {
-      "types": "./dist/types/nations.d.ts",
-      "import": "./dist/types/nations.js",
-      "require": "./dist/types/nations.js"
-    }
-  }
-}
