--- packages/shared/src/index.ts
@@ -1,2 +1,0 @@
-// Re-export all shared types
-export * from './types/nations';
